% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMHN.R
\name{{Modified-Half-Normal}}
\alias{{Modified-Half-Normal}}
\alias{rMHN}
\title{The Modified-Half-Normal Distribution}
\usage{
rMHN(n, alpha, beta, gamma)
}
\arguments{
\item{n}{number of samples to be generated. If \code{length(n) > 1}, the length is taken to
be the number required.}

\item{alpha}{alpha > 0 must be numeric.}

\item{beta}{beta > 0 must be numeric.}

\item{gamma}{gamma is a real number.}
}
\value{
\code{rMHN} generates random deviates.
Invalid arguments will result in error.
The length of the result is determined by \code{n}.
}
\description{
The random number generation for Modified-Half-Normal distribution function
with parameters \code{alpha}, \code{beta}, and \code{gamma}.
}
\details{
The Modified-Half-Normal distribution with parameters \eqn{\alpha} , \eqn{\beta}
and \eqn{\gamma}  has density given by
\deqn{f_{MHN}(x|\alpha,\beta,\gamma) =\frac{2\beta^{\frac{\alpha}{2}}x^{\alpha-1}\exp(-\beta x^2+\gamma x)}{{}_1\Psi_1\begin{bmatrix} \left(\frac{\alpha}{2},\frac{1}{2}\right) ; \frac{\gamma}{\sqrt{\beta}}\\ (1,0) \end{bmatrix}}\mathbb{I}(x>0),}
for \eqn{\alpha > 0, \beta > 0, \gamma \in \mathbb{R}}.
Where  \eqn{\mathbb{I}(x > 0)} denotes the indicator function and \eqn{{}_1\Psi_1\begin{bmatrix} \left(\frac{\alpha}{2},\frac{1}{2}\right) ; z \\ (1,0) \end{bmatrix}}
is a specific case of Fox-Wright Psi function (Fox 1928; Wright 1935).
}
\examples{
ss=rMHN(n=10000,alpha=32,beta=20,gamma=-2)
ss=rMHN(n=1000, alpha=3, beta=1, gamma=1)
ss=rMHN(n=1000, alpha=3, beta=1, gamma=-1)
hist(ss$sample,prob=TRUE, col="grey")
lines(density(ss$sample))
}
\references{
Jingchao Sun, Maiying Kong & Subhadip Pal (2021): The Modified-Half-Normal
distribution: Properties and an efficient sampling scheme, Communications in Statistics - Theory
and Methods.
}
